+++ packages/airless-google-cloud-bigquery/CHANGELOG.md
@@ -1,5 +1,6 @@
 
 **unreleased**
+- [Bugfix] Set default GCP project ID when building bigquery table id
 
 **v0.3.1**
 - [Bugfix] Remove unexisting import for `FileToBigqueryOperator`